feat/storybook Navigation-rail implementation
