+++ apps/@sparrow-storybook/.storybook/preview.js
@@ -1,4 +1,5 @@
 /** @type { import('@storybook/svelte').Preview } */
+import "../../@sparrow-storybook/src/stories/styles/style.scss";
 const preview = {
   actions: { argTypesRegex: "^on.*" },
   parameters: {
