+++ apps/@sparrow-web/src/pages/Dashboard/Dashboard.svelte
@@ -20,7 +20,22 @@
   import { CreateWorkspace } from "@sparrow/teams/features";
   import { CreateTeam } from "@sparrow/common/features";
   import CollectionsPage from "../workspace-page/CollectionsPage.svelte";
-
+  import {
+    home,
+    hoveredHome,
+    selectedHome,
+    settings,
+    selectedSettings,
+    hoveredSettings,
+    collections,
+    hoveredCollections,
+    selectedCollections,
+    help,
+    hoveredHelp,
+    selectedHelp,
+  } from "../../../../../packages/@sparrow-common/src/components/sidebar/common";
+  import { type SidebarItemObj } from "../../../../../packages/@sparrow-common/src/components/sidebar/SidebarItem.svelte";
+  import { isGuestUserActive } from "@app/store/auth.store";
   const _viewModel = new DashboardViewModel();
   let userId;
   const userUnsubscribe = user.subscribe(async (value) => {
@@ -123,6 +138,48 @@
   let showProgressBar = false;
 
   let isCreateTeamModalOpen: boolean = false;
+  isGuestUserActive.subscribe((value) => {
+    isGuestUser = value;
+  });
+
+  let sidebarItems: SidebarItemObj[] = [
+    {
+      route: !isGuestUser ? "/app/home" : "/guest/home",
+      heading: "Home",
+      defaultLogo: home, // Replace with actual image paths
+      hoveredLogo: hoveredHome,
+      selectedLogo: selectedHome,
+      disabled: false,
+      position: "primary",
+    },
+    {
+      route: !isGuestUser ? "/app/collections" : "/guest/collections",
+      heading: "Workspace",
+      defaultLogo: collections,
+      hoveredLogo: hoveredCollections,
+      selectedLogo: selectedCollections,
+      disabled: false,
+      position: "primary",
+    },
+    {
+      route: "/app/help",
+      heading: "Community",
+      defaultLogo: help,
+      hoveredLogo: hoveredHelp,
+      selectedLogo: selectedHelp,
+      disabled: false,
+      position: "secondary",
+    },
+    {
+      route: "/app/setting",
+      heading: "Setting",
+      defaultLogo: settings,
+      hoveredLogo: hoveredSettings,
+      selectedLogo: selectedSettings,
+      disabled: true,
+      position: "secondary",
+    },
+  ];
 </script>
 
 <div class="dashboard d-flex flex-column" style="height: 100vh;">
@@ -190,7 +247,13 @@
     <!-- 
       --Sidebar to naviagte between collection, environment and help page.
     -->
-    <Sidebar {user} onLogout={_viewModel.handleLogout} type="web" />
+    <Sidebar
+      {user}
+      {sidebarItems}
+      {isGuestUser}
+      onLogout={_viewModel.handleLogout}
+      type="web"
+    />
     <!-- 
       -- Dashboard renders any of the pages between collection, environment and help.
     -->
